// src/api/models/select-medication-record-list-by-session-id-res.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface SelectMedicationRecordListBySessionIdRes
 */
export interface SelectMedicationRecordListBySessionIdRes {
    /**
     * 
     * @type {string}
     * @memberof SelectMedicationRecordListBySessionIdRes
     */
    'rowId'?: string;
    /**
     * 
     * @type {string}
     * @memberof SelectMedicationRecordListBySessionIdRes
     */
    'medicationId'?: string;
    /**
     * 
     * @type {string}
     * @memberof SelectMedicationRecordListBySessionIdRes
     */
    'medicationName'?: string;
    /**
     * 
     * @type {number}
     * @memberof SelectMedicationRecordListBySessionIdRes
     */
    'unit'?: number;
    /**
     * 
     * @type {string}
     * @memberof SelectMedicationRecordListBySessionIdRes
     */
    'disposalReason'?: string;
    /**
     * 
     * @type {string}
     * @memberof SelectMedicationRecordListBySessionIdRes
     */
    'createdDatetime'?: string;
    /**
     * 
     * @type {string}
     * @memberof SelectMedicationRecordListBySessionIdRes
     */
    'updatedDatetime'?: string;
    /**
     * 
     * @type {string}
     * @memberof SelectMedicationRecordListBySessionIdRes
     */
    'createdBy'?: string;
    /**
     * 
     * @type {string}
     * @memberof SelectMedicationRecordListBySessionIdRes
     */
    'updatedBy'?: string;
}